move CatalogStatistics into the same file of CatalogTable
