+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/Statistics.scala
@@ -17,12 +17,10 @@
 
 package org.apache.spark.sql.catalyst.plans.logical
 
-import scala.collection.mutable
 import scala.util.control.NonFatal
 
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.{AnalysisException, Row}
-import org.apache.spark.sql.catalyst.catalog.CatalogTable
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.expressions.aggregate._
 import org.apache.spark.sql.types._
@@ -59,37 +57,6 @@ case class Statistics(
     Seq(s"sizeInBytes=$sizeInBytes",
       if (rowCount.isDefined) s"rowCount=${rowCount.get}" else "",
       s"isBroadcastable=$isBroadcastable"
-    ).filter(_.nonEmpty).mkString(", ")
-  }
-}
-
-/**
- * This class of Statistics is used in [[CatalogTable]] to interact with metastore.
- */
-case class CatalogStatistics(
-    sizeInBytes: BigInt,
-    rowCount: Option[BigInt] = None,
-    colStats: Map[String, ColumnStat] = Map.empty) {
-
-  /**
-   * Convert [[CatalogStatistics]] to [[Statistics]], and match column stats to attributes based
-   * on column names.
-   */
-  def convert(attributes: Seq[Attribute]): Statistics = {
-    val matched = mutable.HashMap[Attribute, ColumnStat]()
-    attributes.foreach { attr =>
-      if (colStats.contains(attr.name)) {
-        matched.put(attr, colStats(attr.name))
-      }
-    }
-    Statistics(sizeInBytes = sizeInBytes, rowCount = rowCount,
-      attributeStats = AttributeMap(matched.toSeq))
-  }
-
-  /** Readable string representation for the CatalogStatistics. */
-  def simpleString: String = {
-    Seq(s"sizeInBytes=$sizeInBytes",
-      if (rowCount.isDefined) s"rowCount=${rowCount.get}" else ""
     ).filter(_.nonEmpty).mkString(", ")
   }
 }
